backend/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytesseract==0.3.10
opencv-python==4.9.0.80
pdf2image==1.17.0
Pillow==10.2.0
spacy==3.7.4
nltk==3.8.1
python-dateutil==2.9.0.post0